refactor: extract CanvasWorkspace logic into hooks and sub-components
