+++ apps/web/src/lib/components/canvas/CanvasWorkspace.svelte
@@ -5,15 +5,9 @@
     Background,
     Controls,
     MiniMap,
-    useSvelteFlow,
-    addEdge as addXyEdge,
     ConnectionMode,
-    type Node,
-    type Edge,
-    type Connection,
   } from "@xyflow/svelte";
   import { CanvasStore } from "@codex/canvas-engine";
-  import { vault } from "$lib/stores/vault.svelte";
   import { uiStore } from "$lib/stores/ui.svelte";
   import { canvasRegistry } from "$lib/stores/canvas-registry.svelte";
   import EntityNode from "$lib/components/canvas/EntityNode.svelte";
@@ -22,20 +16,14 @@
   import CustomEdge from "$lib/components/canvas/CustomEdge.svelte";
   import EdgeLabelModal from "$lib/components/canvas/EdgeLabelModal.svelte";
   import CanvasHint from "$lib/components/hints/CanvasHint.svelte";
-  import CategoryFilter from "$lib/components/labels/CategoryFilter.svelte";
+  import CanvasHUD from "./CanvasHUD.svelte";
   import { page } from "$app/state";
-  import { untrack, onDestroy, onMount, tick } from "svelte";
-  import { browser } from "$app/environment";
-  import {
-    buildCanvasSavePayload,
-    createFlowEdgeFromConnection,
-    createFlowEntityNode,
-    hydrateCanvasGraph,
-    pruneCanvasGraph,
-    resolveSpawnPosition,
-  } from "$lib/components/canvas/canvas-workspace-helpers";
+  import { onDestroy } from "svelte";
+  import { createCanvasLogic } from "./use-canvas-logic.svelte";
+  import { useCanvasEvents } from "./use-canvas-events.svelte";
 
   let { engine }: { engine: CanvasStore } = $props();
+  
   const canvasSlug = $derived(page.params.slug);
   const canvas = $derived(
     canvasRegistry.allCanvases.find(
@@ -44,23 +32,19 @@
   );
   const canvasId = $derived(canvas?.id);
 
-  // Category filtering
-  let activeCategories = $state(new Set<string>());
-  function toggleCategoryFilter(categoryId: string) {
-    if (activeCategories.has(categoryId)) {
-      activeCategories.delete(categoryId);
-    } else {
-      activeCategories.add(categoryId);
-    }
-    activeCategories = new Set(activeCategories);
-  }
-  function clearCategoryFilters() {
-    activeCategories = new Set();
-  }
+  const logic = createCanvasLogic(engine);
+  
+  useCanvasEvents({
+    onQuickSpawn: (id, pos, screenPos) => logic.handleQuickSpawn(id, pos, screenPos),
+    onEditLabel: (edgeId, currentLabel) => {
+      logic.labelModal = { isOpen: true, edgeId, currentLabel };
+    },
+    onFlushSave: () => logic.flushSave()
+  });
 
   const filteredNodes = $derived.by(() => {
-    if (activeCategories.size === 0) return nodes;
-    return nodes.filter((n) => activeCategories.has(n.data?.type as string));
+    if (logic.activeCategories.size === 0) return logic.nodes;
+    return logic.nodes.filter((n) => logic.activeCategories.has(n.data?.type as string));
   });
 
   const nodeTypes = {
@@ -72,335 +56,55 @@
     smoothstep: CustomEdge,
   };
 
-  let nodes = $state<Node[]>([]);
-  let edges = $state<Edge[]>([]);
-  let contextMenu = $state<{
-    x: number;
-    y: number;
-    type: "node" | "edge" | "pane";
-    id: string;
-  } | null>(null);
-
-  // Modifier state passed to nodes via context or store
+  // Initialization & Lifecycle
   $effect(() => {
-    const handleDown = (e: KeyboardEvent) => {
-      if (e.key === "Control" || e.metaKey || e.ctrlKey) {
-        uiStore.isModifierPressed = true;
-      }
-    };
-    const handleUp = (e: KeyboardEvent) => {
-      if (e.key === "Control" || e.metaKey || e.ctrlKey) {
-        uiStore.isModifierPressed = false;
-      }
-    };
-    window.addEventListener("keydown", handleDown);
-    window.addEventListener("keyup", handleUp);
-    return () => {
-      window.removeEventListener("keydown", handleDown);
-      window.removeEventListener("keyup", handleUp);
-    };
+    if (canvasId) {
+      logic.initializeCanvas(canvasId);
+    }
   });
 
-  // Label Modal State
-  let labelModal = $state<{
-    isOpen: boolean;
-    edgeId: string;
-    currentLabel: string;
-  }>({
-    isOpen: false,
-    edgeId: "",
-    currentLabel: "",
-  });
-
-  let isConnecting = $state(false);
-
-  let targetVaultId = $state<string | null>(null);
-  let targetCanvasId = $state<string | null>(null);
-  let hasInitialized = $state(false);
-
-  const svelteFlow = useSvelteFlow();
-  const screenToFlowPosition = $derived(svelteFlow?.screenToFlowPosition);
-
+  // Pruning
   $effect(() => {
-    if (browser && engine && canvas) {
-      console.debug("[CanvasWorkspace] Mounting for canvas:", canvas.name, {
-        canvasId: canvas.id,
-        hasSvelteFlow: !!svelteFlow,
-      });
-    }
+    logic.pruneNodes();
   });
 
-  let saveTimer: ReturnType<typeof setTimeout> | null = null;
-  let skipLoadingSaves = 0;
-
-  function flushSave() {
-    if (saveTimer !== null) {
-      clearTimeout(saveTimer);
-      saveTimer = null;
-      untrack(() => {
-        if (targetVaultId && targetCanvasId) {
-          saveCanvas(targetVaultId, targetCanvasId);
-        }
-      });
-    }
-  }
-
-  function debouncedSave() {
-    if (skipLoadingSaves > 0) {
-      skipLoadingSaves--;
-      return;
-    }
-    if (saveTimer !== null) {
-      clearTimeout(saveTimer);
-    }
-    saveTimer = setTimeout(() => {
-      untrack(() => saveCanvas());
-    }, 500);
-  }
-
-  // Ensure registry is loaded for slug resolution (critical for reload/deep-link)
+  // Sync state to engine
   $effect(() => {
-    if (vault.activeVaultId && !canvasRegistry.isLoaded) {
-      untrack(() => canvasRegistry.loadFromVault(vault.activeVaultId!));
-    }
+    logic.syncEngine();
   });
 
-  // Sync engine to local state
+  // Monitor batch spawn
   $effect(() => {
-    if (vault.isInitialized && canvasId && canvasRegistry.isLoaded) {
-      if (targetCanvasId !== canvasId) {
-        // ...
-        untrack(() => {
-          hasInitialized = false;
-          // 1. Flush any pending save for the PREVIOUS canvas before loading new data
-          if (saveTimer !== null && targetVaultId && targetCanvasId) {
-            const oldVaultId = targetVaultId;
-            const oldCanvasId = targetCanvasId;
-            clearTimeout(saveTimer);
-            saveTimer = null;
-            saveCanvas(oldVaultId, oldCanvasId);
-          }
-
-          // 2. Now safe to update the target trackers to the new canvas
-          targetVaultId = vault.activeVaultId;
-          targetCanvasId = canvasId;
-
-          // 3. Load the new data
-          const data = vault.canvases[canvasId];
-
-          if (data) {
-            console.debug("[CanvasWorkspace] Rendering canvas data", canvasId, {
-              nodes: data.nodes?.length,
-              edges: data.edges?.length,
-            });
-
-            // Pre-load all entity contents for the canvas to ensure descriptions/images show up
-            for (const node of data.nodes || []) {
-              vault.loadEntityContent(node.entityId);
-            }
-
-            skipLoadingSaves = 2; // Skip saves triggered by nodes/edges updates
-            const graph = hydrateCanvasGraph(data);
-            nodes = graph.nodes;
-            edges = graph.edges;
-
-            hasInitialized = true;
-            console.debug("[CanvasWorkspace] Canvas mount complete");
-          } else {
-            nodes = [];
-            edges = [];
-            hasInitialized = true;
-          }
-        });
-      }
+    if (canvasRegistry.pendingEntities.length > 0) {
+      logic.handleBatchSpawn();
     }
   });
 
-  // CRITICAL: Monitor vault deletions and remove corresponding nodes from the canvas.
-  // This ensures that deleting an entity in the Graph View also cleans up any active Canvas.
-  $effect(() => {
-    const entityIds = new Set(vault.allEntities.map((e) => e.id));
-    if (hasInitialized && nodes.length > 0) {
-      const pruned = pruneCanvasGraph(nodes, edges, entityIds);
-
-      if (pruned.nodes.length !== nodes.length) {
-        console.debug(
-          `[CanvasWorkspace] Removing ${nodes.length - pruned.nodes.length} nodes due to vault deletion`,
-        );
-        nodes = pruned.nodes;
-        edges = pruned.edges;
-        saveCanvas();
-      }
-    }
-  });
-
-  function onConnect(connection: Connection) {
-    const edgeId = `edge-${crypto.randomUUID()}`;
-    // Explicitly add the edge to our state to ensure reactivity and sync
-    edges = addXyEdge(createFlowEdgeFromConnection(connection, edgeId), edges);
-    // Structural change: save immediately
-    untrack(() => saveCanvas());
-  }
-
-  function onNodeContextMenu({
-    event,
-    node,
-  }: {
-    event: MouseEvent;
-    node: any;
-  }) {
+  function onNodeContextMenu({ event, node }: { event: MouseEvent; node: any }) {
     event.preventDefault();
-    contextMenu = {
-      x: event.clientX,
-      y: event.clientY,
-      type: "node",
-      id: node.id,
-    };
+    logic.contextMenu = { x: event.clientX, y: event.clientY, type: "node", id: node.id };
   }
 
-  function onEdgeContextMenu({
-    event,
-    edge,
-  }: {
-    event: MouseEvent;
-    edge: any;
-  }) {
+  function onEdgeContextMenu({ event, edge }: { event: MouseEvent; edge: any }) {
     event.preventDefault();
-    contextMenu = {
-      x: event.clientX,
-      y: event.clientY,
-      type: "edge",
-      id: edge.id,
-    };
+    logic.contextMenu = { x: event.clientX, y: event.clientY, type: "edge", id: edge.id };
   }
 
   function handlePaneContextMenu({ event }: { event: MouseEvent }) {
     event.preventDefault();
-    contextMenu = {
-      x: event.clientX,
-      y: event.clientY,
-      type: "pane",
-      id: "pane",
-    };
-  }
-
-  async function handleCreateEntity(type: string) {
-    if (!contextMenu) return;
-
-    const title = prompt(`Enter new ${type} name:`);
-    if (!title) return;
-
-    try {
-      const id = await vault.createEntity(type as any, title);
-      const position = screenToFlowPosition({
-        x: contextMenu.x,
-        y: contextMenu.y,
-      });
-
-      const newNodeId = engine.addNode(id, position);
-      nodes = [...nodes, createFlowEntityNode(id, position, newNodeId)];
-      saveCanvas();
-    } catch (err) {
-      console.error("Failed to create entity from canvas", err);
-    }
-  }
-
-  function handleDelete() {
-    if (!contextMenu) return;
-    const targetId = contextMenu.id;
-    if (contextMenu.type === "node") {
-      nodes = nodes.filter((n) => n.id !== targetId);
-      edges = edges.filter(
-        (e) => e.source !== targetId && e.target !== targetId,
-      );
-    } else {
-      edges = edges.filter((e) => e.id !== targetId);
-    }
-    contextMenu = null;
-    // Structural change: save immediately
-    untrack(() => saveCanvas());
-  }
-
-  function handleRename() {
-    if (!contextMenu || contextMenu.type !== "edge") return;
-    const targetId = contextMenu.id;
-    const edge = edges.find((e) => e.id === targetId);
-    labelModal = {
-      isOpen: true,
-      edgeId: targetId,
-      currentLabel: (edge?.label as string) || "",
-    };
-    contextMenu = null;
-  }
-
-  function handleEditLabel(
-    event: CustomEvent<{ edgeId: string; currentLabel: string }>,
-  ) {
-    const { edgeId, currentLabel } = event.detail;
-    labelModal = {
-      isOpen: true,
-      edgeId,
-      currentLabel: currentLabel || "",
-    };
+    logic.contextMenu = { x: event.clientX, y: event.clientY, type: "pane", id: "pane" };
   }
 
   function onEdgeClick({ event, edge }: { event: MouseEvent; edge: any }) {
     if (event.detail === 2) {
-      // Double click
       event.stopPropagation();
-      labelModal = {
+      logic.labelModal = {
         isOpen: true,
         edgeId: edge.id,
         currentLabel: (edge.label as string) || "",
       };
     }
   }
-
-  function saveLabelModal(newLabel: string) {
-    const { edgeId } = labelModal;
-    edges = edges.map((e) => (e.id === edgeId ? { ...e, label: newLabel } : e));
-    saveCanvas();
-  }
-
-  // Keep engine state in sync whenever SvelteFlow's edges change (add/remove).
-  $effect(() => {
-    if (!hasInitialized) return;
-    const snapshot = edges;
-    untrack(() => {
-      if (vault.isInitialized && canvasId) {
-        engine.edges = snapshot.map((e: Edge) => ({
-          id: e.id || `edge-${crypto.randomUUID()}`,
-          source: e.source,
-          target: e.target,
-          sourceHandle: undefined,
-          targetHandle: undefined,
-          label: (e.label as string) || "",
-          type: "straight",
-          style: (e.style as string) || "",
-        }));
-        debouncedSave();
-      }
-    });
-  });
-
-  // Keep engine state in sync whenever SvelteFlow's nodes change (drag/add/remove).
-  $effect(() => {
-    if (!hasInitialized) return;
-    const snapshot = nodes;
-    untrack(() => {
-      if (vault.isInitialized && canvasId) {
-        engine.nodes = snapshot.map((n: Node) => ({
-          id: n.id,
-          type: n.type as "entity",
-          position: n.position,
-          entityId: (n.data?.entityId as string) || "",
-          width: n.data?.width as number,
-          height: n.data?.height as number,
-        }));
-        debouncedSave();
-      }
-    });
-  });
 
   function onDragOver(event: DragEvent) {
     event.preventDefault();
@@ -411,153 +115,20 @@
 
   function onDrop(event: DragEvent) {
     event.preventDefault();
-
     const entityId = event.dataTransfer?.getData("application/codex-entity");
-
     if (!entityId) return;
 
-    const position = screenToFlowPosition({
-      x: event.clientX,
-      y: event.clientY,
-    });
-
-    const newNodeId = engine.addNode(entityId, position);
-    // Ensure the full content (lore, content, image) is loaded for this entity
-    vault.loadEntityContent(entityId);
-
-    // Manually add to nodes to trigger sync
-    nodes = [...nodes, createFlowEntityNode(entityId, position, newNodeId)];
-    saveCanvas();
+    const position = logic.screenToFlowPosition({ x: event.clientX, y: event.clientY });
+    logic.handleQuickSpawn(entityId, position);
   }
-
-  function handleQuickSpawn(
-    event: CustomEvent<{
-      entityId: string;
-      position?: { x: number; y: number };
-      screenPosition?: { x: number; y: number };
-    }>,
-  ) {
-    const {
-      entityId,
-      position: eventPosition,
-      screenPosition: eventScreenPosition,
-    } = event.detail;
-
-    // Prioritize screenPosition (needs conversion), then absolute world position, then center
-    const position =
-      (eventScreenPosition && screenToFlowPosition(eventScreenPosition)) ||
-      eventPosition ||
-      resolveSpawnPosition({
-        screenToFlowPosition,
-        windowSize: {
-          width: window.innerWidth,
-          height: window.innerHeight,
-        },
-      });
-
-    const newNodeId = engine.addNode(entityId, position);
-    // Ensure the full content (lore, content, image) is loaded for this entity
-    vault.loadEntityContent(entityId);
-
-    // Manually add to nodes to trigger sync
-    nodes = [...nodes, createFlowEntityNode(entityId, position, newNodeId)];
-    saveCanvas();
-  }
-
-  function handleBatchSpawn() {
-    const toAdd = canvasRegistry.consumePending();
-    if (toAdd.length === 0) return;
-
-    const newNodesList: Node[] = [];
-
-    toAdd.forEach((item, index) => {
-      const { id: entityId, position: eventPosition } = item;
-
-      // Use provided position (screen coords) or default to staggered center
-      const position = eventPosition
-        ? screenToFlowPosition(eventPosition)
-        : screenToFlowPosition({
-            x: window.innerWidth / 2 + index * 30,
-            y: window.innerHeight / 2 + index * 30,
-          });
-
-      const newNodeId = engine.addNode(entityId, position);
-      vault.loadEntityContent(entityId);
-
-      newNodesList.push({
-        id: newNodeId,
-        type: "entity",
-        position,
-        data: { entityId },
-      });
-    });
-
-    nodes = [...nodes, ...newNodesList];
-    saveCanvas();
-  }
-
-  // Monitor pending entities from registry (e.g. from GraphHUD "Add all results" button)
-  $effect(() => {
-    if (canvasRegistry.pendingEntities.length > 0) {
-      untrack(() => handleBatchSpawn());
-    }
-  });
-
-  async function saveCanvas(
-    explicitVaultId?: string,
-    explicitCanvasId?: string,
-  ) {
-    await tick();
-    const currentVaultId =
-      explicitVaultId || targetVaultId || vault.activeVaultId;
-    const currentCanvasId = explicitCanvasId || targetCanvasId || canvasId;
-
-    if (!currentVaultId || !currentCanvasId) return;
-
-    const exportData = engine.export();
-    const existing = untrack(() => vault.canvases[currentCanvasId] || {});
-
-    // Helper: is the current name just a UUID or "Untitled"?
-    vault.canvases[currentCanvasId] = buildCanvasSavePayload({
-      existing,
-      currentCanvas: canvas,
-      exported: exportData,
-      canvasId: currentCanvasId,
-      lastModified: Date.now(),
-    });
-
-    await vault.saveCanvas(currentCanvasId, {
-      explicitVaultId: currentVaultId,
-    });
-    await canvasRegistry.touch(currentCanvasId);
-  }
-  onMount(() => {
-    window.addEventListener("add-to-canvas", handleQuickSpawn as any);
-    window.addEventListener("edit-edge-label", handleEditLabel as any);
-
-    const handleBeforeUnload = () => flushSave();
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === "hidden") flushSave();
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-    window.addEventListener("visibilitychange", handleVisibilityChange);
-
-    return () => {
-      window.removeEventListener("add-to-canvas", handleQuickSpawn as any);
-      window.removeEventListener("edit-edge-label", handleEditLabel as any);
-      window.removeEventListener("beforeunload", handleBeforeUnload);
-      window.removeEventListener("visibilitychange", handleVisibilityChange);
-    };
-  });
 
   onDestroy(() => {
-    flushSave();
+    logic.flushSave();
   });
 </script>
 
 <div
-  class="canvas-container {isConnecting
+  class="canvas-container {logic.isConnecting
     ? 'is-connecting'
     : ''} flex h-[calc(100vh-var(--header-height,65px))] w-full overflow-hidden relative"
   tabindex="-1"
@@ -570,58 +141,25 @@
     role="region"
     aria-label="Canvas Workspace"
   >
-    <!-- Workspace HUD Overlay -->
-    <div
-      class="absolute top-6 left-6 z-40 flex flex-col items-start gap-2 pointer-events-none select-none"
-    >
-      <button
-        type="button"
-        onclick={() => (uiStore.showCanvasSelector = true)}
-        title="Manage canvases"
-        class="bg-theme-surface/80 backdrop-blur-md border border-theme-primary/30 px-5 py-2 shadow-[0_0_20px_rgba(var(--theme-primary-rgb),0.15)] pointer-events-auto transition-all hover:border-theme-primary/60 group flex items-center gap-2"
-      >
-        <span
-          class="text-xs font-black text-theme-primary uppercase tracking-[0.4em] group-hover:text-theme-accent transition-colors"
-        >
-          {canvas?.name || "Untitled Workspace"}
-        </span>
-        <span
-          class="icon-[lucide--layout-grid] w-3 h-3 text-theme-primary/50 group-hover:text-theme-accent transition-colors shrink-0"
-        ></span>
-      </button>
-
-      <CategoryFilter
-        {activeCategories}
-        onToggle={toggleCategoryFilter}
-        onClear={clearCategoryFilters}
-      />
-
-      {#if canvasRegistry.status === "saving"}
-        <div
-          class="flex items-center gap-2 px-3 py-1 bg-theme-primary/10 border border-theme-primary/20 backdrop-blur-sm animate-pulse"
-        >
-          <span class="icon-[lucide--save] w-3 h-3 text-theme-primary"></span>
-          <span
-            class="text-[8px] font-bold text-theme-primary tracking-[0.2em] uppercase"
-          >
-            Syncing...
-          </span>
-        </div>
-      {/if}
-    </div>
+    <CanvasHUD
+      canvasName={canvas?.name || ""}
+      activeCategories={logic.activeCategories}
+      onToggleCategory={logic.toggleCategoryFilter}
+      onClearCategories={logic.clearCategoryFilters}
+    />
 
     <SvelteFlow
       nodes={filteredNodes}
-      bind:edges
+      bind:edges={logic.edges}
       {nodeTypes}
       {edgeTypes}
-      onconnect={onConnect}
+      onconnect={logic.onConnect}
       onconnectstart={() => {
-        isConnecting = true;
+        logic.isConnecting = true;
         uiStore.isConnecting = true;
       }}
       onconnectend={() => {
-        isConnecting = false;
+        logic.isConnecting = false;
         uiStore.isConnecting = false;
       }}
       onnodecontextmenu={onNodeContextMenu}
@@ -645,15 +183,23 @@
     </SvelteFlow>
   </div>
 
-  {#if contextMenu}
+  {#if logic.contextMenu}
     <CanvasContextMenu
-      x={contextMenu.x}
-      y={contextMenu.y}
-      targetType={contextMenu.type}
-      onDelete={handleDelete}
-      onRename={handleRename}
-      onCreateEntity={handleCreateEntity}
-      onClose={() => (contextMenu = null)}
+      x={logic.contextMenu.x}
+      y={logic.contextMenu.y}
+      targetType={logic.contextMenu.type}
+      onDelete={logic.handleDelete}
+      onRename={() => {
+        const edge = logic.edges.find(e => e.id === logic.contextMenu?.id);
+        logic.labelModal = {
+          isOpen: true,
+          edgeId: logic.contextMenu!.id,
+          currentLabel: (edge?.label as string) || ""
+        };
+        logic.contextMenu = null;
+      }}
+      onCreateEntity={logic.handleCreateEntity}
+      onClose={() => (logic.contextMenu = null)}
     />
   {/if}
 
@@ -661,10 +207,10 @@
   <CanvasSelectionModal />
 
   <EdgeLabelModal
-    bind:isOpen={labelModal.isOpen}
-    initialValue={labelModal.currentLabel}
-    onSave={saveLabelModal}
-    onCancel={() => (labelModal.isOpen = false)}
+    bind:isOpen={logic.labelModal.isOpen}
+    initialValue={logic.labelModal.currentLabel}
+    onSave={logic.saveLabelModal}
+    onCancel={() => (logic.labelModal.isOpen = false)}
   />
 </div>
 
@@ -677,8 +223,6 @@
     background-attachment: fixed;
   }
 
-  /* Force edges to always render strictly underneath all nodes,
-     preventing lines from visibly crossing over a selected node card. */
   :global(.svelte-flow__edges) {
     z-index: 0 !important;
   }
@@ -764,7 +308,6 @@
     fill: var(--color-theme-primary) !important;
     fill-opacity: 0.1 !important;
   }
-  /* High-visibility for the drafting line */
   :global(.svelte-flow__connectionline) {
     z-index: 20 !important;
   }
